op-sync-tester: Reorganize Default EL Sync Policy
